Move data layer logic to shared
