+++ app/src/main/java/io/reark/rxgithubapp/advanced/data/DataLayer.java
@@ -29,9 +29,6 @@ import android.content.Context;
 import android.content.Intent;
 import android.support.annotation.NonNull;
 
-import io.reark.reark.data.DataStreamNotification;
-import io.reark.reark.data.utils.DataLayerUtils;
-import io.reark.reark.pojo.NetworkRequestStatus;
 import io.reark.reark.utils.Preconditions;
 import io.reark.rxgithubapp.advanced.data.stores.GitHubRepositorySearchStore;
 import io.reark.rxgithubapp.advanced.data.stores.GitHubRepositoryStore;
@@ -40,16 +37,10 @@ import io.reark.rxgithubapp.advanced.data.stores.UserSettingsStore;
 import io.reark.rxgithubapp.advanced.network.NetworkService;
 import io.reark.rxgithubapp.shared.data.ClientDataLayerBase;
 import io.reark.rxgithubapp.shared.network.GitHubService;
-import io.reark.rxgithubapp.shared.network.fetchers.GitHubRepositorySearchFetcher;
-import io.reark.rxgithubapp.shared.pojo.GitHubRepository;
-import io.reark.rxgithubapp.shared.pojo.GitHubRepositorySearch;
-import io.reark.rxgithubapp.shared.pojo.UserSettings;
-import rx.Observable;
 
 public class DataLayer extends ClientDataLayerBase {
     private static final String TAG = DataLayer.class.getSimpleName();
     private final Context context;
-    public static final int DEFAULT_USER_ID = 0;
 
     public DataLayer(@NonNull Context context,
                      @NonNull UserSettingsStore userSettingsStore,
@@ -67,68 +58,21 @@ public class DataLayer extends ClientDataLayerBase {
         this.context = context;
     }
 
-    @NonNull
-    public Observable<DataStreamNotification<GitHubRepositorySearch>> getGitHubRepositorySearch(@NonNull final String searchString) {
-        Preconditions.checkNotNull(searchString, "Search string Store cannot be null.");
-
-        final Observable<NetworkRequestStatus> networkRequestStatusObservable =
-                networkRequestStatusStore.getStream(
-                        GitHubRepositorySearchFetcher.getUniqueId(searchString).hashCode());
-        final Observable<GitHubRepositorySearch> gitHubRepositorySearchObservable =
-                gitHubRepositorySearchStore.getStream(searchString);
-        return DataLayerUtils.createDataStreamNotificationObservable(
-                        networkRequestStatusObservable, gitHubRepositorySearchObservable);
-    }
-
-    @NonNull
-    public Observable<DataStreamNotification<GitHubRepositorySearch>> fetchAndGetGitHubRepositorySearch(@NonNull final String searchString) {
-        Preconditions.checkNotNull(searchString, "Search string Store cannot be null.");
-
-        final Observable<DataStreamNotification<GitHubRepositorySearch>> gitHubRepositoryStream =
-                getGitHubRepositorySearch(searchString);
-        fetchGitHubRepositorySearch(searchString);
-        return gitHubRepositoryStream;
-    }
-
-    private void fetchGitHubRepositorySearch(@NonNull final String searchString) {
-        Preconditions.checkNotNull(searchString, "Search string Store cannot be null.");
-
-        Intent intent = new Intent(context, NetworkService.class);
-        intent.putExtra("serviceUriString", GitHubService.REPOSITORY_SEARCH.toString());
-        intent.putExtra("searchString", searchString);
-        context.startService(intent);
-    }
-
-    @NonNull
-    public Observable<GitHubRepository> getGitHubRepository(@NonNull Integer repositoryId) {
-        Preconditions.checkNotNull(repositoryId, "Repository Id cannot be null.");
-
-        return gitHubRepositoryStore.getStream(repositoryId);
-    }
-
-    @NonNull
-    public Observable<GitHubRepository> fetchAndGetGitHubRepository(@NonNull Integer repositoryId) {
-        Preconditions.checkNotNull(repositoryId, "Repository Id cannot be null.");
-
-        fetchGitHubRepository(repositoryId);
-        return getGitHubRepository(repositoryId);
-    }
-
-    private void fetchGitHubRepository(@NonNull Integer repositoryId) {
+    @Override
+    protected void fetchGitHubRepository(@NonNull Integer repositoryId) {
         Intent intent = new Intent(context, NetworkService.class);
         intent.putExtra("serviceUriString", GitHubService.REPOSITORY.toString());
         intent.putExtra("id", repositoryId);
         context.startService(intent);
     }
 
-    @NonNull
-    public Observable<UserSettings> getUserSettings() {
-        return userSettingsStore.getStream(DEFAULT_USER_ID);
-    }
+    @Override
+    protected void fetchGitHubRepositorySearch(@NonNull final String searchString) {
+        Preconditions.checkNotNull(searchString, "Search string Store cannot be null.");
 
-    public void setUserSettings(@NonNull UserSettings userSettings) {
-        Preconditions.checkNotNull(userSettings, "User Settings cannot be null.");
-
-        userSettingsStore.put(userSettings);
+        Intent intent = new Intent(context, NetworkService.class);
+        intent.putExtra("serviceUriString", GitHubService.REPOSITORY_SEARCH.toString());
+        intent.putExtra("searchString", searchString);
+        context.startService(intent);
     }
 }
